Include error handling of runs
Whole runs would crash because of a single failed model. Now there is a "develop" tag in the config file that describes if the program is in development mode (in which case exceptions halt the program) or not. When not then runs continue in a robust manner and will start wi th the next model